:truck:Dividing into files and adding ruff
