+++ .github/workflows/ci.yml
@@ -25,11 +25,15 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-          pip install flake8
+          pip install flake8 ruff
 
       - name: Run linter (flake8)
         run: |
           flake8 src test
+
+      - name: Run linter (ruff)
+        run: |
+          ruff check src test
 
       - name: Run tests
         env:
